Rearrange RSpace serializer files, group in ScodecSerialize object
