+++ rspace/src/main/scala/coop/rchain/rspace/history/HistoryReader.scala
@@ -2,6 +2,7 @@ package coop.rchain.rspace.history
 
 import coop.rchain.rspace.Blake2b256Hash
 import coop.rchain.rspace.internal._
+import coop.rchain.rspace.serializers.ScodecSerialize.{RichDatum, RichJoin, RichKont}
 import scodec.bits.ByteVector
 
 /**
