+++ rspace/src/main/scala/coop/rchain/rspace/state/RSpaceImporter.scala
@@ -4,7 +4,9 @@ import cats.effect._
 import cats.syntax.all._
 import coop.rchain.rspace.Blake2b256Hash
 import coop.rchain.rspace.history._
-import coop.rchain.shared.AttemptOps._
+import coop.rchain.rspace.serializers.ScodecSerialize
+import coop.rchain.rspace.serializers.ScodecSerialize._
+import coop.rchain.shared.AttemptOps.{RichAttempt, _}
 import coop.rchain.shared.Serialize
 import coop.rchain.state.TrieImporter
 import fs2.Stream
@@ -101,7 +103,7 @@ object RSpaceImporter {
                       .delay {
                         persistedData match {
                           case JoinsLeaf(bytes)         => decodeJoins[C](bytes)
-                          case DataLeaf(bytes)          => coop.rchain.rspace.history.decodeData[A](bytes)
+                          case DataLeaf(bytes)          => ScodecSerialize.decodeData[A](bytes)
                           case ContinuationsLeaf(bytes) => decodeContinuations[P, K](bytes)
                         }
                       }
