+++ rspace/src/test/scala/coop/rchain/rspace/ReplayRSpaceTests.scala
@@ -985,7 +985,7 @@ trait ReplayRSpaceTests extends ReplayRSpaceTestsBase[String, Pattern, String, S
       for {
         emptyPoint <- space.createCheckpoint()
 
-        cr = Consume.create(channels, patterns, k, persistent = false)
+        cr = Consume(channels, patterns, k, persistent = false)
 
         _ <- consumeMany(
               space,
