+++ rspace/src/test/scala/coop/rchain/rspace/history/EncodingSpec.scala
@@ -4,6 +4,7 @@ import coop.rchain.rspace.Blake2b256Hash
 import coop.rchain.rspace.examples.StringExamples.implicits._
 import coop.rchain.rspace.examples.StringExamples.{Pattern, StringsCaptor}
 import coop.rchain.rspace.internal.{Datum, _}
+import coop.rchain.rspace.serializers.ScodecSerialize._
 import coop.rchain.rspace.test.ArbitraryInstances._
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
