+++ rspace/src/test/scala/coop/rchain/rspace/trace/EventTests.scala
@@ -1,28 +1,21 @@
 package coop.rchain.rspace.trace
 
-import cats.implicits._
+import coop.rchain.rspace.Blake2b256Hash
 import coop.rchain.rspace.StableHashProvider._
 import coop.rchain.rspace.examples.StringExamples.implicits._
 import coop.rchain.rspace.examples.StringExamples.{Pattern, StringsCaptor}
-import coop.rchain.rspace.internal._
+import coop.rchain.rspace.serializers.ScodecSerialize.{RichAttempt, _}
 import coop.rchain.rspace.test.ArbitraryInstances._
-import coop.rchain.rspace.Blake2b256Hash
+import coop.rchain.shared.Serialize
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
-import scodec.bits.{BitVector, ByteVector}
 import scodec.codecs.{ignore => cignore, _}
-import scodec.interop.cats._
-import scodec.{Attempt, Codec}
-import coop.rchain.rspace.util
-import coop.rchain.shared.Serialize
-
-import scala.collection.immutable.Seq
 
 class EventTests extends FlatSpec with Matchers with GeneratorDrivenPropertyChecks {
 
   "A Produce" should "contain the expected hash" in {
     forAll { (channel: String, data: String, persist: Boolean) =>
-      val actual = Produce.create(channel, data, persist)
+      val actual = Produce(channel, data, persist)
 
       val expectedHash =
         Blake2b256Hash.create(
@@ -44,7 +37,7 @@ class EventTests extends FlatSpec with Matchers with GeneratorDrivenPropertyChec
 
       val continuation = new StringsCaptor
 
-      val actual = Consume.create(channels, patterns, continuation, persist)
+      val actual = Consume(channels, patterns, continuation, persist)
 
       val encodedChannels = toOrderedByteVectors(channels)
       val encodedPatterns = toOrderedByteVectors(patterns)
@@ -68,8 +61,8 @@ class EventTests extends FlatSpec with Matchers with GeneratorDrivenPropertyChec
       val (channels, patterns) = channelPatterns.unzip
       val continuation         = new StringsCaptor
 
-      val actual   = Consume.create(channels, patterns, continuation, persist)
-      val reversed = Consume.create(channels.reverse, patterns.reverse, continuation, persist)
+      val actual   = Consume(channels, patterns, continuation, persist)
+      val reversed = Consume(channels.reverse, patterns.reverse, continuation, persist)
 
       reversed.hash shouldBe actual.hash
     }
